### 2024-04-22 #6659 @editorialbot
[PRE REVIEW]: GOBLIN Lite: A National Land Balance Model for Assessment of Climate Mitigation Pathways for Ireland
**Submitting author:** @colmduff (<a href="http://orcid.org/0000-0002-0076-6749">Colm Duffy</a>)
**Repository:** https://github.com/GOBLIN-Proj/goblin_lite
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.4
