### 2024-04-22 #6659 @kthyng
Is there a reason users need to install from github instead of setting up a release with PyPI or conda-forge?
