### 2024-04-22 #6659 @colmduff
Hi!

I haven't actually put it on either of those. Caveating my submission here with the fact that I am a bit of a "learn as you go" developer. I can put it on those platforms, but honestly, had not investigated at this point.

In terms of install instructions, which looks to be the only thing you need right now, I had just put pip install instructions. But, I can have a look through other repos that have published in JOSS and confirm my documents.

Apologies again if you have to tell me any seemingly basic things, its just a little inexperienced. So all criticism is most welcome.

Thanks a million

Colm
