### 2024-04-22 #6659 @kthyng
That is ok! JOSS wants to be available to submitting authors like you describe yourself as with less experience, but we also are willing to point you to resources to up your game, as it were. In this case, I suggest you use this time (you're going to be on the waitlist anyway though you didn't know it yet) to get your repo onto PyPI since it's the easiest one, and conda-forge if possible. Here are some links to help:
* https://www.pyopensci.org/python-package-guide/
* https://www.pyopensci.org/python-package-guide/package-structure-code/publish-python-package-pypi-conda.html#
