### 2024-04-23 #6659 @colmduff
OK, perfect.

I will get some reviewers for you from the db later today, update the docs with install instructions, and get working on PyPI.

Very much appreciate the guidance and the patience.

Thanks a million!

Colm
