### 2024-05-07 #6659 @mengqi-z
> Checking the BibTeX entries failed with the following error:
>
> ```
> Failed to parse BibTeX on value "b" (NAME) ["@", #<BibTeX::Entry >, {:url=>["https://gmd.copernicus.org/articles/15/2239/2022/"], :archiveprefix=>["gmd"], :author=>["{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D."], :journal=>["Geoscientific Model Development"], :month=>[:mar], :title=>["{GOBLIN version 1.0: a land balance model to identify national agriculture and land use pathways to climate neutrality via backcasting}"], :year=>"2022"}]
> ```

Hi @colmduff, I'll be your editor for this submission. While I am finding some reviewers, would you mind addressing the following errors in your `.bib` file?
