+++ joss/paper.bib
@@ -5,7 +5,7 @@
   	Journal = {Geoscientific Model Development},
   	Month = mar,
   	Title = {{GOBLIN version 1.0: a land balance model to identify national agriculture and land use pathways to climate neutrality via backcasting}},
-  	Year = 2022b
+  	Year = 2022
 }
 @article{Kurz:2008,
   	url = {https://www.sciencedirect.com/science/article/abs/pii/S0304380008005012},
@@ -28,10 +28,10 @@
   	Author = {{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {Iannetta}, P.P.M. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D.},
   	Journal = {Nature Sustainability},
   	Title = {{Randomized national land management strategies for net-zero emissions}},
-  	Year = 2022a
+  	Year = 2022
 }
 @misc{NIR,
-  author = {Duffy, P. and Black, K. and Fahey, D. and Hyde, B. and Kehoe, A. and Monaghan, S. and Murphy, J. and Ryan, A.M. and Ponzi, J},
+  author = {Duffy, P. and Black, K. and Fahey, D. and Hyde, B. and Kehoe, A. and Monaghan, S. and Murphy, J. and Ryan, A.M. and Ponzi, J.},
   title = {IRELAND’S NATIONAL INVENTORY REPORT 2022: GREENHOUSE GAS EMISSIONS 1990 - 2020 REPORTED TO THE UNITED NATIONS FRAMEWORK CONVENTION ON CLIMATE CHANGE},
   year = {2022},
   publisher = {Environmental Protection Agency, Ireland},
