+++ tests/directory_test.py
@@ -1,15 +1,27 @@
 import os 
 from goblin_lite.resource_manager.directory import Directories
+import shutil
 
 def main():
 
     os.makedirs("./data/directory_test/", exist_ok=True)
 
     # class instances
-    dir_class = Directories("./data/directory_test/", gen_baseline=True, gen_validation=False)
+    
 
     # create directory structure
-    dir_class.create_goblin_directory_strucutre(10)
+    for i in range(10):
+        os.mkdir(f"./data/directory_test/{i}")
+        path = os.path.join("./data/directory_test/", str(i)+"/")
+
+        print(f"Creating directory structure for {path}")
+        dir_class = Directories(path)
+        dir_class.create_goblin_directory_structure()
+
+    # delete the directory
+    shutil.rmtree("./data/directory_test/")
+
+    print("Directory test passed.")
 
 if __name__ == "__main__":
     main()
