+++ INSTALLATION.md
@@ -1,0 +1,44 @@
+Installation
+============
+
+This section provides instructions on how to install the software.
+
+Requirements
+------------
+The software requires the following:
+
+- Python 3.9 or higher
+- pandas 2.1.4
+- numpy 1.25.0 or higher
+- sqlalchemy-utils latest
+- cattle_lca 0.3.1
+- sheep_lca 0.3.1
+- crop_lca 0.2.1
+- grassland_production 0.3.3
+- livestock_generation 0.2.1
+- static_scenario_generator 1.1.1
+- landcover_assignment 0.2.1
+- landcover_lca 0.1.2
+- goblin_cbm_runner 0.3.4
+- scenario_assessment 0.1.3
+- seaborn = 0.13.1 or higher
+- matplotlib latest
+
+Install with pip
+----------------
+You can install the package using pip:
+
+```bash
+
+    pip install goblin_lite
+```
+
+Install from Source
+-------------------
+To install from source using pip and git:
+
+```bash
+    pip install git+https://github.com/GOBLIN-Proj/goblin_lite.git@main
+```
+
+
