### 2024-05-08 ab1808c
update docs, and requirements
