### 2024-05-08 #6659 @mengqi-z
> > > Checking the BibTeX entries failed with the following error:
> > > ```
> > > Failed to parse BibTeX on value "b" (NAME) ["@", #<BibTeX::Entry >, {:url=>["https://gmd.copernicus.org/articles/15/2239/2022/"], :archiveprefix=>["gmd"], :author=>["{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D."], :journal=>["Geoscientific Model Development"], :month=>[:mar], :title=>["{GOBLIN version 1.0: a land balance model to identify national agriculture and land use pathways to climate neutrality via backcasting}"], :year=>"2022"}]
> > > ```
> >
> >
> > Hi @colmduff, I'll be your editor for this submission. While I am finding some reviewers, would you mind addressing the following errors in your `.bib` file?
>
> Hi @mengqi-z
>
> Just looking at that error. I have tried to address that, but I think its the "b" on the end of the year, which I need to distinguish between two similar citations. It seems to compile fine for me at the moment. But, in saying that, it also compiled for me before as well.

Hi @colmduff,

I believe you can remove `a` and `b` and keep the year as `2022` for both papers because it should automatically distinguish between the papers when it gets compiled.
