### 2024-05-08 #6732 @mengqi-z
@colmduff, The reason it didn't work could be because the author lists from two papers are not exactly the same (In your case, the first four authors are the same, but the rest are not). You could try adding the [`key` field](https://bib-it.sourceforge.net/help/fieldsAndEntryTypes.php#key) in both entries to distinguish between the papers. There should be a way to address this.
