+++ joss/paper.bib
@@ -3,10 +3,8 @@
   	Archiveprefix = {gmd},
   	Author = {{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D.},
   	Journal = {Geoscientific Model Development},
-  	Month = mar,
   	Title = {{GOBLIN version 1.0: a land balance model to identify national agriculture and land use pathways to climate neutrality via backcasting}},
-  	Year = 2022,
-    key = {DuffyB2022}
+  	Year = 2022b
 }
 @article{Kurz:2008,
   	url = {https://www.sciencedirect.com/science/article/abs/pii/S0304380008005012},
@@ -29,8 +27,7 @@
   	Author = {{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {Iannetta}, P.P.M. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D.},
   	Journal = {Nature Sustainability},
   	Title = {{Randomized national land management strategies for net-zero emissions}},
-  	Year = 2022,
-    key = {DuffyA2022}
+  	Year = 2022a
 }
 @misc{NIR,
   author = {Duffy, P. and Black, K. and Fahey, D. and Hyde, B. and Kehoe, A. and Monaghan, S. and Murphy, J. and Ryan, A.M. and Ponzi, J.},
