+++ joss/paper.bib
@@ -4,7 +4,7 @@
   	Author = {{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D.},
   	Journal = {Geoscientific Model Development},
   	Title = {{GOBLIN version 1.0: a land balance model to identify national agriculture and land use pathways to climate neutrality via backcasting}},
-  	Year = 2022b
+  	Year = 2022
 }
 @article{Kurz:2008,
   	url = {https://www.sciencedirect.com/science/article/abs/pii/S0304380008005012},
@@ -27,7 +27,7 @@
   	Author = {{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {Iannetta}, P.P.M. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D.},
   	Journal = {Nature Sustainability},
   	Title = {{Randomized national land management strategies for net-zero emissions}},
-  	Year = 2022a
+  	Year = 2022
 }
 @misc{NIR,
   author = {Duffy, P. and Black, K. and Fahey, D. and Hyde, B. and Kehoe, A. and Monaghan, S. and Murphy, J. and Ryan, A.M. and Ponzi, J.},
