### 2024-06-12 #6732 @david-yannick
I am complete with my review, here are a few questions/comments I have regarding the paper

Paper.md
-	Overall, software paper is concise and well written.
-	Based on previous work cited utilizing the original GOBLIN model [lines 24-25], which also has been used in several studies in the AFOLU sector [lines 30-32]. This is specialized to work in this particular need.
-	Are calculated fluxes similar to a micrometeorological/atmospheric sign convention? Ecosystem CO2 uptake/photosynthesis a negative flux, and fluxes released to atmosphere are positive. It may be helpful to state.
-	Figure 2 – Croplands not represented in example? It is mentioned that there is a reduction in livestock being converted to forest [50-51]. Were livestock lands considered croplands in this example or perhaps not labeled/represented in figure?
-	Line 21 CH4 formatting needs to be addressed
