### 2024-06-18 #6732 @colmduff
Hi @david-yannick

Thanks a million for your time an effort on this. Apologies for my late response, I have had some revision deadlines on another paper.

To answer your questions:

- calculated fluxes: Yes, emission == positive, and removal == negative
- Figure 2: Croplands actually are in there, they are just super small. Crop area in ireland is < 5% of total agricultural area, so the sector is very heavily dominated by livestock production.
- Formatting: Good catch, my brain may have filtered that out when I was proof reading.

Thanks again for your time and effort on this, it is greatly appreciated.
