### 2024-06-19 #6732 @varsha2509
Hello. I've completed my review and here are my comments. Overall, well written paper and code is well structured. Proposed some comments/suggestions below to improve code readability.
Comments on paper:

- Line 21 - typo after CH4, remove _
- Line 48  - Looking at the CMB-CSF3 paper, the authors mention that the flexible nature of the model has enabled it to be adapted for use in several other countries. Are there any limitations that users should be aware of for the use of this model within Irish context? Came across this tool (https://www.teagasc.ie/crops/forestry/advice/environment/forest-carbon-tool/) that may be specific to IE. Is this something the authors have considered?
- Figure 2 -Can the authors add more description in the caption explaining each of those figures? Can the authors also please add a little bit more description on what Scenario 0 and 1 indicate? I see the information provided in config.json, but what do these two scenarios indicate?
- Figure 3 - The figure is quite small and hard to read. Can the authors make the figure larger and provide more information on what the results mean? For instance, under the baseline, we see an increase in carbon removal by above ground biomass between ~2030 - 2040, but carbon removal by ABG reduces overtime for scenario 0 and 1. Why is that?
- Recommend renaming scenarios to base something like Baseilne (scenario 0), low upland sheep population or something along those lines for scenario 1, high upland sheep population for scenario 2.
- Figure 4 - please enlarge the figure and include more description about the results (even if the results are for illustrative purposes).
- Figure 4 - Scenarios are mapped as scenario 0, 1, 2  - these should be -1, 0 and 1?
- In lines 67 - 72, authors mentioned that care must be taken while defining inputs. What checks/bounds have the authors added to the model in order to validate the inputs and outputs produced? If this tool is designed for use in policy decisions, having some validation can go a long way. Is this something authors will consider in future iterations of the tool?
-Can authors please include the limitations of this work so that users of this tool are aware of limitations?

Minor comments on code:

- goblin.py - There seems to be some redundant code here specifically [lines 111 - 115](https://github.com/GOBLIN-Proj/goblin_lite/blob/main/src/goblin_lite/goblin.py#L111). The reassignment of instance variables to local variables is not necessary here. When referring to instance variables in run_scenarios, authors can use `self.` to access them instead.

The authors should consider breaking down `run_scenarios` into smaller modules to improve readability. For instance, the authors can make use of the `ScenarioRunner` class and its methods more efficiently:
The code for animal data generator could be simplified to look like:

```
baseline_animal_data, scenario_animal_data, protein_and_milk_summary = self._generate_animal_data(scenario_input_dataframe)

def _generate_animal_data(self, scenario_input_dataframe):
        animal_data_generator = AnimalDataGenerator(
            self.ef_country, self.calibration_year, self.target_year, scenario_input_dataframe
        )
        baseline_animal_data, scenario_animal_data = animal_data_generator.generate_animal_data()
        protein_and_milk_summary = animal_data_generator.generate_livestock_outputs()
        return baseline_animal_data, scenario_animal_data, protein_and_milk_summary
```

LCA_processing.py - similar comment as above, there seems to be some redundant code in some of the .py files that can be removed.

Sorry for the delay in getting my comments to you. Let me know if you have questions.
