+++ pyproject.toml
@@ -17,8 +17,8 @@ crop_lca = "^0.2.1"
 grassland_production = "^0.3.0"
 livestock_generation = "^0.2.1"
 static_scenario_generator = "^1.1.1"
-landcover_assignment  = "^0.2.0"
-landcover_lca  = "^0.1.2"
+landcover_assignment = { git = "https://github.com/GOBLIN-Proj/landcover_assignment.git", branch = "EFs" }
+landcover_lca = { git = "https://github.com/GOBLIN-Proj/landcover_lca.git", branch = "EFs" }
 goblin_cbm_runner = "^0.3.4"
 scenario_assessment = "^0.1.3"
 seaborn = "^0.13.1"
