### 2024-06-20 #6732 @colmduff
Hi @varsha2509

Thanks a million for your detailed feedback. I will try to repond to the comments here, I will post the combined changes I have made in a seperate post before I resubmit the paper.

To make things a little easier to follow, I am just going to number the response to correspond to the order in which you gave feedback:

1. Typo fixed
2. Detailed parameterisation of the CBM CFS3 is discussed in the cited national inventory report, and is beyond the scope of this paper. The model development here is aimed at aligning with national inventory reporting, which is partly why we have integrated the model here. The Teagasc model, as far as I am aware, is a relatively simple stand level model in excel format.
3. Greater detail added to the caption. These scenarios are arbitrary and meant to illustrate the type of outputs available. I have cited numerous scenario level papers and methodological background to point the reader towards additional resources. I feel that evolving this snap shot overview of the model into more than that is potentially beyond the scope of this paper.
4. I have made several adjustments here. I have combined everything into a single graphic, which will make things a little bigger. I have also added some contextual clarity in terms of assumptions made with regard to afforestation in the baseline scenario. I have also pointed the user toward additional materials.
5. We have, in previous work, by default, labeled the baseline scenario as (-1). I will keep this naming format for consistency.
6. Figure enlarged, additional text added.
7. Typo in figure corrected
8. In terms of input validation, there is a minimum level at the moment. We are also working with the Irish Centre for High End Computing, implementing an online simplified version for public use. This will have an online dashboard for inputs, and additional validation will be prioritised as part of the development. At the moment, the tool probably has significant barriers to entry (ability to code) for policymakers. It can assist in generating output to assist in policy decision making, but the average policy maker is unlikley to engage directly.
9. Added limitations section
10. This is discussed in cited peer reviewed articles, and is, in my opinion, perhaps beyond the scope here.

Minor comments:

Thank you for taking the time to go through the code in such detail, it is very much appreciated.

I will look to stream line some functions further in the future, this is an ongoing process. With regards to redundant code, I do realise that there is not a specific need to have these variables here. But, it is a personal choice on my part, just to be explicit and seeing variables I am dealing with right in front of me just helps me organise my thoughts a little better. I realise this may offend a developer here and there, but, I am willing to run that risk if it helps me with my workflow.

Thank you again for all your hard work on this, your feedback is very much appreciated.
