+++ joss/paper.md
@@ -36,7 +36,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-Ireland's complex AFOLU emissions profile requires a modeling tool that effectively captures both major greenhouse gas (GHG) sources and sinks. `GOBLIN Lite` is designed to meet this need. Based on the original GOBLIN (General Overview for a Backcasting approach to Livestock Intensification) model [@DuffyA:2022], it calculates CO~2~ fluxes from varied soil types and forests, along with CH~4_ emissions from livestock activities. It also models nitrogen (N) losses, including N~2~O, NH~3~, and dissolved forms, across the agricultural landscape.
+Ireland's complex AFOLU emissions profile requires a modeling tool that effectively captures both major greenhouse gas (GHG) sources and sinks. `GOBLIN Lite` is designed to meet this need. Based on the original GOBLIN (General Overview for a Backcasting approach to Livestock Intensification) model [@DuffyA:2022], it calculates CO~2~ fluxes from varied soil types and forests, along with CH~4~ emissions from livestock activities. It also models nitrogen (N) losses, including N~2~O, NH~3~, and dissolved forms, across the agricultural landscape.
 
 `GOBLIN Lite`'s enhanced resolution – in herd dynamics, soil group-specific spared area modeling, and additional impact categories – surpasses the original GOBLIN model [@DuffyA:2022]. It integrates Tier 3 forest modeling [@Kurz:2008] for increased detail and aligns with Irish National Inventory reporting [@NIR]. These features make `GOBLIN Lite` a powerful asset for policymakers, researchers, and students, as evidenced by its use in impactful "net-zero" pathway studies.
 
@@ -46,7 +46,7 @@ The GOBLIN modeling framework has already been used in recent studies on pathway
 
 Figure 1 illustrates the model architecture and data flow. The scenario module generates model input parameters based on user data (a JSON or CSV input). These outputs are used by the livestock module to determine national livestock (cattle and sheep) numbers for each scenario. The grassland production module then calculates the area required to support these herds/flocks based on livestock energy requirements (from the cattle and sheep LCA modules) and nitrogen (N) response curves [@McEniry:2013]. The N response curve takes into account organic and inorganic fertilizer inputs, as well as soil group. System-specific utilization rates are derived from National Farm Survey data [@NFS]. The total spared area available for alternative land uses is the difference in grassland required between the scenario and the baseline. Final spared areas are disaggregated by soil group. The land use module assigns spared area and produces a transition matrix. Spared area assigned to afforestation is mapped to forest productivity classes [@Farrelly:2015].
 
-![GOBLIN Lite Architecture and Data Flow.\label{fig:Figure2}](goblin_data_flow.drawio.png)
+![GOBLIN Lite Architecture and Data Flow.\label{fig:Figure1}](goblin_data_flow.drawio.png)
 
 Life Cycle Assessment (LCA) processing is handled directly within `GOBLIN Lite`. The package coordinates the flow of data and contains the impact categories. For forest emissions and removals, the CBM-CSF3 Python library is used.
 
@@ -54,15 +54,15 @@ Life Cycle Assessment (LCA) processing is handled directly within `GOBLIN Lite`.
 
[…]
