+++ src/goblin_lite/scenario_analysis/data_grapher.py
@@ -136,7 +136,9 @@ class DataGrapher:
 
         frames = self.get_combined_crop_and_livestock_pathway_frames()
 
-        xlabels = [label for label in frames[0].index]
+        print(frames)
+
+        xlabels = [label -1 for label in frames[0].index]
 
         sns.set()
         sns.set_style("white")
