### 2024-06-20 #6732 @colmduff
Hi @mengqi-z @varsha2509 @david-yannick

Thank you all for your help and engagement with this. Its very much appreciated.

I have responded to each reviewer individually, I just want to summarise the changes to the document here in a consolidated way.

- Typo in line 21 corrected
- Fig 2 has been edited to include additional detail. Additional supporting detail also added.
- Fig 3 has also been edit to produce a single, larger figure, with additional supporting detail added
- Fig 4, additional detail added. However, I have tried to enlarge this figure, but it seems to be at the maximum allowable. This is perhaps due to the width of the figure itself. It is still readable at a moderate zoom level. I can look at rendering an additional row of subplots, rather than a single row. Perhaps that might help, if the current format is not feasible.
- Typo in Figure 4 corrected
- Limitations section has been added.
