### 2024-07-08 7ef97f0
structure and bugs
