+++ docs/conf.py
@@ -9,7 +9,7 @@
 project = "GOBLIN lite"
 copyright = "2024, Colm Duffy"
 author = "Colm Duffy"
-release = "0.4.0"
+release = "0.4.1"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
