+++ src/goblin_lite/impact_categories/climate_change.py
@@ -1,10 +1,10 @@
 """
-Impact Categories Module
-=========================
+Impact Category Climate Change
+==============================
 
-The Impact Categories module is designed to calculate and analyze various environmental impacts in the context of land use change, livestock and crop production. 
-This includes emissions related to climate change, eutrophication, and air quality. The module integrates data from various sources like cattle 
-and sheep lifecycle assessments, crop production data, and land use changes, providing a comprehensive view of environmental impacts.
+The Impact Categories Climate Change module is designed to calculate and analyze various environmental impacts in the context of land use change, livestock and crop production. 
+The module integrates data from various sources like cattle and sheep lifecycle assessments, crop production data, 
+and land use changes, providing a comprehensive view of environmental impacts.
 
 Key Features
 ------------
@@ -18,25 +18,13 @@ Flexible Data Handling: Works with different types of data inputs, including liv
 from cattle_lca.resource_manager.models import load_livestock_data, load_farm_data
 from cattle_lca.lca import ClimateChangeTotals as CattleClimateChangeTotals
 from sheep_lca.lca import ClimateChangeTotals as SheepClimateChangeTotals
-from cattle_lca.lca import EutrophicationTotals as CattleEutrophicationTotals
-from sheep_lca.lca import EutrophicationTotals as SheepEutrophicationTotals
-from cattle_lca.lca import AirQualityTotals as CattleAirQualityTotals
-from sheep_lca.lca import AirQualityTotals as SheepAirQualityTotals
 from crop_lca.models import load_crop_farm_data
 from crop_lca.lca import ClimateChangeTotals as CropClimateChangeTotals
-from crop_lca.lca import EutrophicationTotals as CropEutrophicationTotals
-from crop_lca.lca import AirQualityTotals as CropAirQualityTotals
 import landcover_lca.lca_emission as landuse_lca
 import landcover_lca.models as landuse_models
 from goblin_lite.resource_manager.goblin_data_manager import GoblinDataManager
 import pandas as pd
-
-
-class CommonParams:
-    def __init__(self) -> None:
-        self.baseline_index = -1
-        self.kg_to_kt = 1e-6
-        self.t_to_kt = 1e-3
+from goblin_lite.impact_categories.common import CommonParams
 
 
 class ClimateChangeLandUse:
@@ -116,7 +104,8 @@ class ClimateChangeLandUse:
         )
 
         emission_df.index.levels[0].astype(int)
-        past_forest_mask = (forest_data["Year"] == baseline) & (
+        #need to add one year to the forest baseline year to get the correct data
+        past_forest_mask = (forest_data["Year"] == baseline +1) & (
             forest_data["Scenario"] == baseline_index
         )
 
@@ -150,9 +139,6 @@ class ClimateChangeLandUse:
                     land_use_data[base],
                     transition_matrix[base],
                     ef_country,
-                )
-                + landuse_lca.horticulture_co2_peat_export(
-                    ef_country, baseline, baseline
                 )
             )
             * kg_to_kt
@@ -286,7 +272,6 @@ class ClimateChangeLandUse:
                 transition_matrix[base],
                 ef_country,
             )
-            + landuse_lca.horticulture_co2_peat_export(ef_country, baseline, baseline)
         ) * kg_to_kt
 
         emission_df.loc[
@@ -443,9 +428,6 @@ class ClimateChangeLandUse:
                         transition_matrix[sc],
                         ef_country,
                     )
-                    + landuse_lca.horticulture_co2_peat_export(
-                        ef_country, target_year, baseline
-                    )
                 )
                 * kg_to_kt
             ) + (
@@ -577,9 +559,6 @@ class ClimateChangeLandUse:
                     land_use_data[base],
                     transition_matrix[sc],
[…]
