+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "goblin_lite"
-version = "0.4.0"
+version = "0.4.1"
 description = "A goblin tool for the generation of static land balance scenarios and environmental impacts"
 authors = ["Colm Duffy"]
 license = "GPL-3.0"
@@ -19,7 +19,7 @@ livestock_generation = "^0.2.1"
 static_scenario_generator = "^1.1.1"
 landcover_assignment = "^0.3.0"
 landcover_lca = "^0.2.0"
-goblin_cbm_runner = "^0.3.4"
+goblin_cbm_runner = "^0.3.7"
 scenario_assessment = "^0.1.3"
 seaborn = "^0.13.1"
 matplotlib = "*"

[…]
