### 2024-07-08 #6732 @colmduff
Hi @mengqi-z, @varsha2509, @david-yannick. I hope you are all well. We have been working on the National Inventory updates for several weeks now. The updated emissions factors and land area classifications for 2024 have a significant impact on organic soils under grassland, and wetland emissions. Essentially, the reclassification of grassland means that emissions there are reduced given a lower emissions factor and the fact that a large proportion is now classed as naturally rewetted rather than drained.

Regarding wetlands, revised emisson factors here mean that wetlands are a significantly higher source, as newly incorporated research indicates that no wetland is in a natural state, all have been impacted to by athropogenic activity to some degree.

Regeneration of the data using the updated National Inventory emissions factors and classifications for 2024 are the only changes made.

I have added some national level emissions outputs below. For land cover emisisons and agricutlure emissions. Inputs are generated for the goblin model based on 2024 inputs, but these are not exact, some aggregation, with regards to fertiliser etc. is required. But the outputs serve to illustrate GOBLIN's capacity to replicate the 2024 national inventory.

![landuse_emission_validation](https://github.com/openjournals/joss-reviews/assets/88794612/7ed99aa4-55ed-4927-be12-0056896a3c70)
![livestock_emission_validation](https://github.com/openjournals/joss-reviews/assets/88794612/03faed80-82f9-41d6-b1f8-c94f8e8a471a)
