### 2024-07-23 #6732 @mengqi-z
@david-yannick @varsha2509 Thank you both for your constructive review!

@colmduff I will make my pass on the paper over the next few days and provide next steps.
