### 2024-08-04 GOBLIN-Proj/goblin_lite#49 @mengqi-z
JOSS Paper Review
Hi @colmduff,

There are few additional issues to address in the paper for JOSS https://github.com/openjournals/joss-reviews/issues/6732

1. Line 38: Please define "LCA" the first time it appears in the paper.
2. Figure 3: Could you double-check if the y-axis label unit "t/Carbon" is correct for carbon flux?
3. Please correct the missing DOIs in the References mentioned below.

- For the first 4 references listed below: Ensure to use DOI links instead of website links for articles that have a DOI. For example, for the Duffy et al., (2022) GMD paper, use the DOI "https://doi.org/10.5194/gmd-15-2239-2022" instead of "https://gmd.copernicus.org/articles/15/2239/2022/"
- For references numbered 5 to 9 listed below: Verify if they have DOIs. If not, please provide the URL instead. For example, if there is no DOI for McEniry et al., (2013), use the URL "http://www.jstor.org/stable/23631018".

```

OK DOIs

- None

MISSING DOIs

- 10.5194/gmd-15-2239-2022 may be a valid DOI for title: GOBLIN version 1.0: a land balance model to identi...
- 10.1016/j.ecolmodel.2008.10.018 may be a valid DOI for title: CBM-CFS3: A model of carbon-dynamics in forestry a...
- 10.1038/s43247-024-01275-0 may be a valid DOI for title: Defining national net zero goals is critical for f...
- 10.1038/s41893-022-00946-0 may be a valid DOI for title: Randomized national land management strategies for...
- No DOI given, and none found for title: IRELAND’S NATIONAL INVENTORY REPORT 2022: GREENHOU...
- No DOI given, and none found for title: How much grassland biomass is available in Ireland...
- No DOI given, and none found for title: Teagasc National Farm Survey 2021
- No DOI given, and none found for title: The potential availability of land for afforestati...
- No DOI given, and none found for title: GOBLIN Lite

INVALID DOIs

- None
```
