### 2024-08-04 #6732 @mengqi-z
@colmduff I have proofread the paper and create an issue (https://github.com/GOBLIN-Proj/goblin_lite/issues/49) for some remaining items for you to address in the paper.

After you address my comments, could you please:

- [ ] Double check authors and affiliations (including ORCIDs).
- [ ] Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).
- [ ] Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.
- [ ] Please post the DOI of the archived version here.

Once these steps are completed, I can move forward with accepting the submission.
