+++ joss/paper.bib
@@ -1,6 +1,6 @@
 
 @article{DuffyB:2022,
-  	url = {https://gmd.copernicus.org/articles/15/2239/2022/},
+  	doi = {10.5194/gmd-15-2239-2022},
   	Archiveprefix = {gmd},
   	Author = {{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D.},
   	Journal = {Geoscientific Model Development},
@@ -8,7 +8,7 @@
   	Year = 2022
 }
 @article{Kurz:2008,
-  	url = {https://www.sciencedirect.com/science/article/abs/pii/S0304380008005012},
+  	doi = {10.1016/j.ecolmodel.2008.10.018},
   	Archiveprefix = {ecolmodel},
   	Author = {{Kurz}, W.A. and {Dymond}, C.C. and {White}, T.M. and {Stinson}, G. and {Shaw}, C.H. and {Rampley}, G.J. and {Smith}, C. and {Simpson}, B.N. and {Neilson}, E.T. and {Trofymow}, J.A. and {Metsaranta}, J. and {Apps}, M.J.},
   	Journal = {Ecological Modelling},
@@ -17,14 +17,14 @@
   	Year = 2008
 }
 @article{Bishop:2024,
-  	url = {https://doi.org/10.1038/s43247-024-01275-0},
+  	doi = {10.1038/s43247-024-01275-0},
   	Author = {{Bishop}, G. and {Duffy}, C. and {Prudhomme}, R. and {Cowie}, A. and {O'Donoghue}, C. and {Cain}, M. and {Lanigan}, G.J. and {Styles}, D.},
   	Journal = {Communications Earth \& Environment},
   	Title = {{Defining national net zero goals is critical for food and land use policy}},
   	Year = 2024
 }
 @article{DuffyA:2022,
-  	url = {https://doi.org/10.1038/s41893-022-00946-0},
+  	doi = {10.1038/s41893-022-00946-0},
   	Author = {{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {Iannetta}, P.P.M. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D.},
   	Journal = {Nature Sustainability},
   	Title = {{Randomized national land management strategies for net-zero emissions}},
@@ -43,6 +43,7 @@
   title = {How much grassland biomass is available in Ireland in excess of livestock requirements?},
   journal = {Irish Journal of Agricultural and Food Research},
   year = {2013}
+  url ={http://www.jstor.org/stable/23631018}
 }
 @misc{NFS,
   author = {Dillon, E. and Donnellan, T. and Moran, B. and Lennon, J.},
