+++ joss/paper.bib
@@ -42,7 +42,7 @@
   author = {McEniry, J. and Crosson, P. and Finneran, E. and Mcgee, M. and Keady, T.W.J. and O’kiely, P.},
   title = {How much grassland biomass is available in Ireland in excess of livestock requirements?},
   journal = {Irish Journal of Agricultural and Food Research},
-  year = {2013}
+  year = {2013},
   url ={http://www.jstor.org/stable/23631018}
 }
 @misc{NFS,
