### 2024-08-06 #6732 @colmduff
> @colmduff I have proofread the paper and create an issue ([GOBLIN-Proj/goblin_lite#49](https://github.com/GOBLIN-Proj/goblin_lite/issues/49)) for some remaining items for you to address in the paper.
>
> After you address my comments, could you please:
>
> * [x]  Double check authors and affiliations (including ORCIDs).
> * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [x]  Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).
> * [x]  Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.
> * [x]  Make sure that the license listed for the archive is the same as the software license.
> * [x]  Please post the DOI of the archived version here.
>
> Once these steps are completed, I can move forward with accepting the submission.

@mengqi-z, thanks for this. I have addressed the issues above, I will make the text edits now.

The files are in "pending" status on zenodo, but I have the doi: 10.5281/zenodo.13239928

The version uploaded to PyPi is: v0.4.2
