### 2024-08-06 GOBLIN-Proj/goblin_lite#49 @colmduff
[x] Line 38: Please define "LCA" the first time it appears in the paper.
[x] Figure 3: Could you double-check if the y-axis label unit "t/Carbon" is correct for carbon flux?
[x] Please correct the missing DOIs in the References mentioned below.

Figure 3, for additional clarity, the outputs from libcbm are initially in tonnes of carbon. They are converted to CO2e later, but the raw times lines in the figure are directly from libcbm.
