### 2024-08-07 #6732 @colmduff
@mengqi-z, the issue with zenodo is resolved (needed a zip).
 [goblin_lite_0.4.2](https://zenodo.org/records/13239928?token=eyJhbGciOiJIUzUxMiJ9.eyJpZCI6ImM2MDAwMzgzLTQyN2EtNDk4YS04MDMxLTJhMGMxYzMwMmVlOCIsImRhdGEiOnt9LCJyYW5kb20iOiI5OGQ1YTIxZTEzNjRkNDYxZjkxYjkyNWNkM2U5MjFmMCJ9.qlqnbIb_g6jwQdpHw7SRboi-XjOFM1yQd1Fs_sP9DGQT1Io8tinfaSm7pEu3qIFbEnX6KAQkmShJ1F1BQPFMrA)
