### 2024-08-09 #6732 @mengqi-z
@colmduff - Could you please update the Zenodo metadata to ensure that all authors are listed in the author section, matching the author list in your paper? You can keep the current contributor roles as well if you prefer. Thanks!
