### 2024-08-12 #6732 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
