### 2024-08-15 #6732 @kthyng
@colmduff
* looks like you need a tag or release on github to match your Zenodo and JOSS version unless I am missing something — the last tag I see is one behind.
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "ireland" is not capitalized two places and maybe that one reference doesn't need to be capitalized the whole way through?
