+++ joss/paper.bib
@@ -39,7 +39,7 @@
   url = {https://www.epa.ie/publications/monitoring--assessment/climate-change/air-emissions/Ireland-NIR-2022_Merge_v2..pdf}
 }
 @article{McEniry:2013,
-  author = {McEniry, J. and Crosson, P. and Finneran, E. and Mcgee, M. and Keady, T.W.J. and O’kiely, P.},
+  author = {McEniry, J. and Crosson, P. and Finneran, E. and Mcgee, M. and Keady, T.W.J. and O’Kiely, P.},
   title = {How much grassland biomass is available in Ireland in excess of livestock requirements?},
   journal = {Irish Journal of Agricultural and Food Research},
   year = {2013},
