+++ joss/paper.bib
@@ -13,26 +13,26 @@
   	Author = {{Kurz}, W.A. and {Dymond}, C.C. and {White}, T.M. and {Stinson}, G. and {Shaw}, C.H. and {Rampley}, G.J. and {Smith}, C. and {Simpson}, B.N. and {Neilson}, E.T. and {Trofymow}, J.A. and {Metsaranta}, J. and {Apps}, M.J.},
   	Journal = {Ecological Modelling},
   	Month = mar,
-  	Title = {{CBM-CFS3: A model of carbon-dynamics in forestry and land-use change implementing IPCC standards}},
+  	Title = {{CBM-CFS3: A model of carbon-dynamics in forestry and land-use change implementing {IPCC} standards}},
   	Year = 2008
 }
 @article{Bishop:2024,
   	doi = {10.1038/s43247-024-01275-0},
   	Author = {{Bishop}, G. and {Duffy}, C. and {Prudhomme}, R. and {Cowie}, A. and {O'Donoghue}, C. and {Cain}, M. and {Lanigan}, G.J. and {Styles}, D.},
   	Journal = {Communications Earth \& Environment},
-  	Title = {{Defining national net zero goals is critical for food and land use policy}},
+  	Title = {{Defining national {Net-Zero} goals is critical for food and land use policy}},
   	Year = 2024
 }
 @article{DuffyA:2022,
   	doi = {10.1038/s41893-022-00946-0},
   	Author = {{Duffy}, C. and {Prudhomme}, R. and {Duffy}, B. and {Gibbons}, J. and {Iannetta}, P.P.M. and {O'Donoghue}, C. and {Ryan}, M. and {Styles}, D.},
   	Journal = {Nature Sustainability},
-  	Title = {{Randomized national land management strategies for net-zero emissions}},
+  	Title = {{Randomized national land management strategies for {Net-Zero} emissions}},
   	Year = 2022
 }
 @misc{NIR,
   author = {Duffy, P. and Black, K. and Fahey, D. and Hyde, B. and Kehoe, A. and Monaghan, S. and Murphy, J. and Ryan, A.M. and Ponzi, J.},
-  title = {IRELAND’S NATIONAL INVENTORY REPORT 2022: GREENHOUSE GAS EMISSIONS 1990 - 2020 REPORTED TO THE UNITED NATIONS FRAMEWORK CONVENTION ON CLIMATE CHANGE},
+  title = {{Ireland's National Inventory Report 2022: Greenhouse Gas Emissions 1990 - 2020 Reported to the United Nations Framework Convention on Climate Change}},
   year = {2022},
   publisher = {Environmental Protection Agency, Ireland},
   journal = {National Inventory Report},
@@ -40,28 +40,28 @@
 }
 @article{McEniry:2013,
   author = {McEniry, J. and Crosson, P. and Finneran, E. and Mcgee, M. and Keady, T.W.J. and O’Kiely, P.},
-  title = {How much grassland biomass is available in Ireland in excess of livestock requirements?},
+  title = {How much grassland biomass is available in {Ireland} in excess of livestock requirements?},
   journal = {Irish Journal of Agricultural and Food Research},
   year = {2013},
   url ={http://www.jstor.org/stable/23631018}
 }
 @misc{NFS,
   author = {Dillon, E. and Donnellan, T. and Moran, B. and Lennon, J.},
-  title = {Teagasc National Farm Survey 2021},
+  title = {{Teagasc National Farm Survey 2021}},
   year = {2022},
   journal = {National Farm Survey},
   url = {https://www.teagasc.ie/media/website/publications/2022/Teagasc-National-Farm-Survey-2021.pdf}
 }
 @article{Farrelly:2015,
   author = {Farrelly, N. and Gallagher, G.},
[…]
