### 2024-08-22 #6732 @kthyng
Congratulations on your new publication @colmduff! Many thanks to @mengqi-z and to reviewers @david-yannick and @varsha2509 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.

@colmduff If you're interested in joining JOSS as a reviewer, please sign up at https://reviewers.joss.theoj.org/!
